requirements.txt
djangorestframework-simplejwt
djangorestframework
pasycopg2
django-seed
pillow
django
